Switch to sending data instead of saving it in a JSON file
I can't easily get iCloud to serve a hosted version of the JSON file, it
seems you can only get a link to a html page with a download button. I
think it makes sense to then just send the data to trmnl (with
debouncing so that we're not hitting the server so frequently!)

This patch starts to move away from saving the JSON file, but it's not
entirely complete. Have added a setting to store the URL we want to hit,
but need to perform the fetch, remove the file saving and do some cleanup :)